LPS-105884 Move package
